Assert that previously significant run stays significant
In some cases, runs which have been deemed significant, based on
OMS data that RR receives, suddenly become non-significant on run
update, due to data that has been removed or changed from the OMS
side. This should not, in principle, happen.

This commit adds such an assertion to throw an Exception
if this case arises. Without it, the code continues, and in most
cases reaches the point where RR tries to delete data from its
database (usually LS info), raising an error (`Lumisections cannot
be deleted`). The assertion catches this a bit earlier,
stopping with a more appropriate message.

This is probably not ideal, and this check could be placed even
earlier.