[WIP] Add placeholder LS info if OMS does not have it
It has been observed that sometimes, the OMS API returns
LS information with some of them missing, e.g., a run might
have 10 LS but OMS returns 8, with some missing in-between.
Usually, those fall in the range of LS that have
`cms_active=false`, but this is not always the case.

This commit adds some placeholder LS information.